moved app code to src and packaged w setuptools
