+++ src/allocation/adapters/orm.py
@@ -2,7 +2,7 @@ from sqlalchemy import (
     MetaData, Table, Column, Integer, String, Date, ForeignKey
 )
 from sqlalchemy.orm import mapper, relationship
-from model import OrderLine, Batch
+from allocation.domain import model
 
 # https://docs.sqlalchemy.org/en/13/core/metadata.html#sqlalchemy.schema.MetaData
 # Holds a collection of Table objects and optional binding to
@@ -34,12 +34,12 @@ allocations = Table(
 )
 
 def start_mappers():
-    lines_mapper = mapper(OrderLine, order_lines)  # returns Mapper object that defines correlation
+    lines_mapper = mapper(model.OrderLine, order_lines)  # returns Mapper object that defines correlation
     # of class attrs to ddbb table columns. When mapper() is used explicitly to link a user defined
     # class with table metadata, this is referred to as classical mapping.
     # https://docs.sqlalchemy.org/en/13/orm/mapping_api.html#sqlalchemy.orm.mapper.params.properties
     # https://docs.sqlalchemy.org/en/13/orm/relationship_api.html#sqlalchemy.orm.relationship
-    mapper(Batch, batches, properties={
+    mapper(model.Batch, batches, properties={
         '_allocations': relationship(
             lines_mapper,  # mapped class or Mapper instance representing relationship target
             secondary=allocations, # intermediary junction table to link two tables
