+++ src/allocation/domain/model.py
